dating_django/requirements.txt
asgiref==3.5.0
Django==3.2.12
django-crispy-forms==1.14.0
django-filter==21.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
install==1.3.5
Pillow==9.0.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.1.1
whitenoise==6.0.0
